cleaned up tests to use new database based classes
